renamed samplers => sampler and a few documentation tweaks
