+++ src/data/data.js
@@ -1,11 +1,11 @@
 import feed from './feed/feed';
 import random from './random/random';
-import samplers from './samplers/samplers';
+import sampler from './sampler/sampler';
 import spread from './spread';
 
 export default {
     feed: feed,
     random: random,
     spread: spread,
-    samplers: samplers
+    sampler: sampler
 };
