+++ src/data/sampler/modeMedian.js
